tabular/requirements.txt
pandas==2.0.1
hydra_core==1.3.2
omegaconf==2.3.0
wandb==0.15.2
python_dotenv==1.0.0